Update run-system-tests.pl
Line 6 doesn't work, but there are two possible reasons (and possible fixes). Either
A) the test should be looking for the build decoder in $cdec/build/decoder/cdec instead of in $cdec/decoder/cdec.
or
B) the "make -j4" command should be creating a decoder executable at $cdec/decoder/cdec , but isn't.

I assumed (A), because it's simpler.